Update url for getting started with slack webhooks
The currently linked implementation leads to an error "invalid_token" when trying to use with the plugin. 
The approach that must be followed for creation of the Slack Webhook needs to be as a Slack App. 
Hence, I swapped out the link here so future new starters set up with the correct Webhook and aren't stuck in the same rabbithole as I was.

Ref:https://discourse.drone.io/t/invalid-token-error-message-when-trying-to-use-slack-notifications/8391/2